Refactor seller draft and listing components for improved loading state management
- Integrated `Suspense` and `withFallback` into `DraftList`, `ListingList`, and `Item` components to enhance loading experiences.
- Removed redundant `Pending` and `Data` components to streamline the structure and improve maintainability.
- Updated various components to utilize colocated fallback UIs, ensuring consistent loading states across the application.
- Enhanced user interaction by adding `data-action` attributes for better accessibility and feedback.
