+++ apps/app/src/app/@seller/draft/DraftListPage/DraftList/DraftList.tsx
@@ -1,9 +1,15 @@
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { Container, SpinnerContainer } from "@use-pico/client/ui/container";
+import { EmptyState } from "@use-pico/client/ui/empty-state";
+import { withFallback } from "@use-pico/client/utils";
+import { withDraftQuery } from "@zbav-se.me/sdk/query/seller/draft";
+import { Suspense, useMemo } from "react";
+import { CreateButton } from "../../CreateButton";
+import { Empty } from "./Empty";
+import { Item } from "./Item";
 
 export namespace DraftList {
-	export interface Props extends Omit<Data.Props, "_suspense"> {
+	export interface Props extends Container.Props, MarkSuspense.Props {
 		//
 	}
 }
@@ -14,13 +20,61 @@ export namespace DraftList {
  *
  * @see apps/app/src/app//draft/page/DraftListPage.tsx
  */
-export const DraftList: FC<DraftList.Props> = (props) => {
+export const DraftList = withFallback(({ _suspense, ui, ...props }: DraftList.Props) => {
+	const { data: draftCollection } = withDraftQuery.useCollectionQuery({
+		where: {
+			usedAtIsNull: true,
+		},
+		sort: [
+			{
+				field: "updatedAt",
+				order: "desc",
+			},
+		],
+	});
+
+	const check = useMemo(() => {
+		return [
+			{
+				check() {
+					return !draftCollection.length;
+				},
+				render() {
+					return <Empty />;
+				},
+			},
+		] as EmptyState.Check[];
+	}, [
+		draftCollection,
+	]);
+
 	return (
-		<Suspense fallback={<Pending />}>
-			<Data
-				_suspense={"I know"}
-				{...props}
-			/>
-		</Suspense>
+		<Container
+			data-ui="DraftList[Container]"
+			ui={{
+				scroll: "vertical",
+				height: "full",
+				layout: "vertical-flex",
+				gap: "default",
+				...ui,
+			}}
+			{...props}
+		>
+			<EmptyState check={check}>
+				{draftCollection.map((draftId) => (
+					<Suspense
+						key={draftId}
+						fallback={<Item.Fallback />}
+					>
+						<Item
+							_suspense={"I know"}
+							draftId={draftId}
+						/>
+					</Suspense>
+				))}
+
+				<CreateButton />
+			</EmptyState>
+		</Container>
 	);
-};
+}, SpinnerContainer);
