--- apps/app/src/app/@seller/draft/DraftListPage/DraftList/Item/Pending.tsx
@@ -1,18 +1,0 @@
-import { SpinnerContainer } from "@use-pico/client/ui/container";
-import type { FC } from "react";
-
-export namespace Pending {
-	export interface Props extends SpinnerContainer.Props {
-		//
-	}
-}
-
-export const Pending: FC<Pending.Props> = (props) => {
-	return (
-		<SpinnerContainer
-			data-ui="DraftList-[SpinnerContainer]"
-			type="icon"
-			{...props}
-		/>
-	);
-};
