+++ apps/app/src/app/@seller/listing/MyListingPage/ListingList/Content.tsx
@@ -1,9 +1,8 @@
 import type { MarkSuspense } from "@use-pico/client/type";
+import { Suspense, type FC, useCallback } from "react";
 import { VisibleContainer } from "@use-pico/client/ui/container";
-import { type FC, useCallback } from "react";
 import { CreateButton } from "~/app/@seller/draft/~public/CreateButton";
 import { ListingItem } from "./ListingItem";
-import { Pending } from "./ListingItem/Pending";
 
 export namespace Content {
 	export interface Props extends MarkSuspense.Props {
@@ -13,7 +12,7 @@ export namespace Content {
 
 export const Content: FC<Content.Props> = ({ _suspense, listingIds }) => {
 	const placeholder = useCallback(() => {
-		return <Pending />;
+		return <ListingItem.Fallback />;
 	}, []);
 
 	return (
@@ -25,10 +24,9 @@ export const Content: FC<Content.Props> = ({ _suspense, listingIds }) => {
 						id={listingId}
 						placeholder={placeholder}
 					>
-						<ListingItem
-							key={listingId}
-							listingId={listingId}
-						/>
+						<Suspense fallback={<ListingItem.Fallback />}>
+							<ListingItem listingId={listingId} />
+						</Suspense>
 					</VisibleContainer>
 				);
 			})}
