--- apps/app/src/app/@seller/listing/MyListingPage/ListingList/Data/index.ts
@@ -1,1 +1,0 @@
-export { Data } from "./Data";
